RCA_Docker/requirements.txt
numpy==1.15.4
torch==1.5.0
torchvision==0.6.0
pandas==0.23.4
scikit-image==0.16.2
Pillow==5.3.0
dominate==2.3.5   
pathlib
